Merge upstream: v0.12.2 → v0.12.3 (#3)
* Fix `;` character in URI OSC 8 payload

The special character `;` can be not URL-encoded, thus it'll add
extra parameter in the payload. Handle it joining extra parameters
with the `;` as a separator.

* Fix selection rotation on the last line

This fixes an issue with terminal resizes when the selection is on the
last line. Alacritty would fail to rotate lines and keep the selection
in the same line index whenever the terminal line count was grown or
shrunk.

This issue occurred due to the range passed to the selection's rotate
function still being based on the old terminal size, which caused the
initial or target state of the rotation to be outside of the terminal
bounds.

Closes #6698.

* Bump winit to 0.28.4

Fixes #6817.

* Add `ALACRITTY_EXTRA_LOG_TARGETS` env variable

This should help with troubleshooting the dependency crates like winit.

* Bump winit to 0.28.5

The 0.28.4 was yanked.

* Alacritty version 0.12.1-rc1

* Fix very long startup time on some Wayland systems

This is not a real fix for the issue given that dbus method sctk-adwaita
is using will being called anyway. The proper fix will be available
with the winit's 0.29.0 release.

Right now the delay reduced from around 20 seconds to 100ms on a
systems with broken xdg-desktop-portal setup.

* Update glutin to 0.30.8 and winit to 0.28.6

Fixes #6920.
Fixes #6824.

* Alacritty version 0.12.1-rc2

* Bump `actions/checkout` to v3

Fix Node 12 deprecation warnings in action runs.

* Fix upload from CI on windows

* Alacritty version v0.12.1

* Fix hyperlink preview for 2 lines terminal (#6953)

The intention was to show it, however it was hidden due to wrong
comparisson check.

* Fix crash on ScaleFactorChange on Windows

Windows is known to send zero sizes from winit in Risezed and now
in ScaleFactorChanged events. They were handled in Resized, but
not in ScaleFactorChanged.

Fixes #6949.

* Fix freeze with some drivers when using GLX

Fixes #6972.

* Fix the crash when shrinking scrolled terminal

display_offset was adjusted unconditionally, thus it could go
beyound the history limits, so clamp it to history like we do
in grow_colums.

Fixes #6862.

* Update wix on the release CI

Fixes #6939.

* Alacritty version v0.12.2-rc1

* Alacritty version v0.12.2

* Update winit to 0.28.7 and sctk to 0.16.1

This resolves crashes on macOS Sonoma and Wayland compositors
advertising newer `wl_compositor` versions.

* Alacritty version 0.12.3-rc1

* Alacritty version 0.12.3

---------

Co-authored-by: Kirill Chibisov <contact@kchibisov.com>
Co-authored-by: Christian Duerr <contact@christianduerr.com>
Co-authored-by: K.B.Dharun Krishna <kbdharunkrishna@gmail.com>